Update Book.md
I suspect pull requests aren't the best way to give editorial feedback. For example, "this" at the beginning of the piece should be replaced, but it is unclear to me what it should be replaced with - so in that case, I would leave a comment or a few suggestions.  

I also don't understand "a day 2 overview" - i suspect that's a sociolect or a technolect. Depending on the target audience of this book, it might be fine, or it might benefit from an edit. 

I cleared up some language and fixed a few typos / grammar issues.